Use .github/release_type for controlling releases of the OSS/EE [DI-52] (#204)
In order to control which editions (OSS/EE/ALL) should be released
`.github/release_type` file is introduced.

This file is required now for newly created git tags. It should contains
only one of the following values:
- OSS
- EE
- ALL

When a new tag is pushed we read the file to decide which editions
should be released.
For manual runs of the `publish-packages.yml` we ignore this file and
use `RELEASE_TYPE` input from the workflow

`.github/release_type` should be added during the release by the
automation. It's forbidden to add it master or maintenance branches
through PRs.

Fixes: https://hazelcast.atlassian.net/browse/DI-52

---------

Co-authored-by: Jack Green <jack.green@hazelcast.com>