Fix `Install Hazelcast from deb`'s `apt install` [DI-358] (#245)
`apt install` prompts for confirmation, which caused the job to fail
when not specified - with the newly added option, this is avoided.

Note that I do not understand why this is only _now_ causing a problem,
as recent changes do not explain it:
- https://github.com/hazelcast/hazelcast-packaging/pull/220
   - I can reproduce this issue with both `apt-get` and `apt`
- https://github.com/hazelcast/hazelcast-packaging/pull/241
   - I can reproduce this issue with both Ubuntu and Debian

I think it should _always_ have failed.

Fixes: [DI-358](https://hazelcast.atlassian.net/browse/DI-358)

[DI-358]:
https://hazelcast.atlassian.net/browse/DI-358?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ