Make Java version dynamic (#278)
We _should_ only need to specify the Java version once and populate it
in the various package configurations dynamically.

Note - in the case of Homebrew (at least), the Java version change is
the _only_ difference between `5.3.z` & `master` - i.e. if this was
dynamic we could use the same `rb` template for all versions.