Fix PR builder not triggering `oss` `job`s (#291)
In https://github.com/hazelcast/hazelcast-packaging/pull/273
(specifically -
https://github.com/hazelcast/hazelcast-packaging/pull/273#discussion_r2398709569)
the logic for determining which edition to build was updated.

The assumption was that boolean logic can be used, but because of truthy
logic, the result was unexpected. More specifically, the `deb-oss` `job`
was never triggered, blocking PR merges.


https://github.com/hazelcast/hazelcast-packaging/blob/822c36448414af71c5fd0fa608d3997f1574404a/.github/workflows/publish-packages.yml#L53

`resolve-editions.outputs.should_build_oss ` returns `yes` or `no`, but
_any_ result would be considered `true` in this `if`


https://github.com/hazelcast/hazelcast-packaging/blob/822c36448414af71c5fd0fa608d3997f1574404a/.github/workflows/publish-packages.yml#L125

`prepare.outputs.should_build_oss ` is the result of a boolean
expression, it will be `true` or `false`, but never `yes`.

Reworked and simplified.