Externalise Hazelcast Packaging _source_ artefacts location [DI-650] (#283)
To support sandbox, the source repository(s) of the artefacts to be
packaged needs to be dynamic - and support for private repositories is
required too.

This was already achieved in `hazelcast-docker` with:
- https://github.com/hazelcast/hazelcast-docker/pull/1111
- https://github.com/hazelcast/hazelcast-docker/pull/1116

For consistency, implemented _mostly_ the same way:
- matching repository definitions added to the environment
- repository definitions that require authentication tagged with an ID
(e.g. `my-private-repo::::https://example.com/private-repo`) - does not
need to be secret
- authentication credentials supplied from a secret
- download artefacts using Maven _where possible_ to allow joining
credentials to repositories to authenticate
- where not possible (Homebrew), is skipped completely

Post-merge actions:
- [ ] _Consider_ consolidating distribution download between
packaging/docker

Fixes: [DI-650](https://hazelcast.atlassian.net/browse/DI-650)

[DI-650]:
https://hazelcast.atlassian.net/browse/DI-650?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ