Fix homebrew on sandbox (#311)
[Execution
fails](https://github.com/hz-devops-test/hazelcast-packaging/actions/runs/20273543068/job/58215741303)
with:
> Could not transfer artifact {...} status code: 401, reason phrase:
(401)

This is because the Maven authentication was missed from the brew action
- but features everywhere else.
This wasn't spotted earlier because [private artifacts are not
supported](https://hazelcast.atlassian.net/browse/DI-659), so a failure
was expected - but a _subsequent_ step was supposed to identify this and
print a more helpful error than this.

I have _not_ tested this in sandbox, but can if required.